Finish Building View Refactor to Cards and Dockerfile changes

+++ src/main/java/kyo/yaz/condominium/manager/persistence/entity/Building.java
@@ -62,5 +62,5 @@ public class Building {
     @JsonProperty
     private final BigDecimal reserveFundPercentage;
     @JsonProperty
-    private final Integer amountOfApts;
+    private final Long amountOfApts;
 }
